requirement.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2
django-crispy-forms==1.11.2
django-formtools==2.3
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
